Simplify dependence on esm-framework in app shell
