+++ packages/framework/esm-globals/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./events";
-export * from "./globals";
 export * from "./types";
