+++ packages/shell/esm-app-shell/src/run.ts
@@ -1,9 +1,3 @@
-import "import-map-overrides";
-import "systemjs/dist/system";
-import "systemjs/dist/extras/amd";
-import "systemjs/dist/extras/named-exports";
-import "systemjs/dist/extras/named-register";
-import "systemjs/dist/extras/use-default";
 import { start, triggerAppChange } from "single-spa";
 import {
   setupApiModule,
